requirements.txt
flet
langchain
langchain-google-genai
python-dotenv
pydantic
